Kafka added on Event Service

+++ event-service/src/main/java/com/example/event_service/dto/EventResponse.java
@@ -7,7 +7,7 @@ import java.time.LocalDateTime;
 
 @Data
 public class EventResponse {
-    private int eventId;
+    private Long eventId;
     private String title;
     private String description;
     private String category;
